Split the message schemas out of the main FMN package
