--- .coveragerc
@@ -1,24 +1,0 @@
-# SPDX-FileCopyrightText: Contributors to the Fedora Project
-#
-# SPDX-License-Identifier: MIT
-
-[run]
-branch = True
-source = fmn
-omit =
-    fmn/database/migrations/env.py
-    fmn/database/migrations/versions/*.py
-concurrency =
-    greenlet
-    thread
-
-[report]
-precision = 2
-fail_under = 100
-exclude_lines =
-    pragma: no cover
-    def __repr__
-    if TYPE_CHECKING:
-    raise NotImplementedError
-    \.\.\.
-show_missing = True
