+++ .reuse/dep5
@@ -11,6 +11,6 @@ Files: .github/*.md
 Copyright: Contributors to the Fedora Project
 License: MIT
 
-Files: poetry.lock .github/renovate.json
+Files: poetry.lock fmn-messages/poetry.lock .github/renovate.json
 Copyright: None, autogenerated
 License: MIT
