+++ changelog.d/+splitted-schemas.changed
@@ -1,0 +1,1 @@
+The message schemas are now in a separate fmn-messages packages, that can be installed independently
