+++ fmn-messages/README.md
@@ -1,0 +1,15 @@
+<!--
+SPDX-FileCopyrightText: Contributors to the Fedora Project
+
+SPDX-License-Identifier: MIT
+-->
+
+# FMN messages
+
+A schema package for [FMN](http://github.com/fedora-infra/fmn).
+
+See the [detailed documentation](https://fedora-messaging.readthedocs.io/en/latest/messages.html) on packaging your schemas.
+
+![PyPI](https://img.shields.io/pypi/v/fmn-messages.svg)
+![Supported Python versions](https://img.shields.io/pypi/pyversions/fmn-messages.svg)
+![Build status](http://github.com/fedora-infra/fmn-messages/actions/workflows/main.yml/badge.svg?branch=main)
