+++ fmn/api/handlers/users.py
@@ -11,10 +11,11 @@ from sqlalchemy import func, select
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from fmn_messages.rule import RuleCreateV1, RuleDeleteV1, RuleUpdateV1
+
 from ...backends import FASJSONAsyncProxy, get_fasjson_proxy
 from ...core.constants import DEFAULT_MATRIX_DOMAIN
 from ...database.model import Destination, Filter, Generated, GenerationRule, Rule, User
-from ...messages.rule import RuleCreateV1, RuleDeleteV1, RuleUpdateV1
 from .. import api_models
 from ..auth import Identity, get_identity, get_identity_optional
 from ..database import gen_db_session
