+++ tox.ini
@@ -17,15 +17,15 @@ allowlist_externals =
 commands_pre =
     poetry install --all-extras
 commands =
-  poetry run pytest -o 'addopts=--cov-config .coveragerc --cov=fmn --cov-report term-missing --cov-report xml --cov-report html' --asyncio-mode auto tests/ {posargs}
+  poetry run pytest -o 'addopts=--cov=fmn --cov-report term-missing --cov-report xml --cov-report html' --asyncio-mode auto tests/ {posargs}
 
 [testenv:black]
 commands =
-  poetry run black --diff --check fmn/ tests/
+  poetry run black --diff --check fmn/ tests/ fmn-messages/fmn_messages/
 
 [testenv:lint]
 commands =
-  poetry run ruff check fmn/ tests/
+  poetry run ruff check --diff fmn/ tests/ fmn-messages/fmn_messages/
   poetry run reuse lint
 
 [testenv:docs]
